+++ relm4/src/channel/shutdown/receiver.rs
@@ -1,8 +1,11 @@
 // Copyright 2022 System76 <info@system76.com>
 // SPDX-License-Identifier: MIT or Apache-2.0
 
+use super::broadcast::Receiver;
 use super::AttachedShutdown;
-use async_broadcast::Receiver;
+
+#[cfg(test)]
+use super::broadcast::error::TryRecvError;
 
 /// Listens to shutdown signals and constructs shutdown futures.
 #[derive(Debug)]
@@ -25,15 +28,15 @@ impl ShutdownReceiver {
     }
 
     #[cfg(test)]
-    pub(crate) fn try_recv(&mut self) -> Option<()> {
-        self.receiver.try_recv().ok()
+    pub(crate) fn try_recv(&mut self) -> Result<(), TryRecvError> {
+        self.receiver.try_recv()
     }
 }
 
 impl Clone for ShutdownReceiver {
     fn clone(&self) -> Self {
         Self {
-            receiver: self.receiver.clone(),
+            receiver: self.receiver.resubscribe(),
         }
     }
 }