+++ relm4/src/component/sync/message_broker.rs
@@ -53,27 +53,30 @@ impl<C: Component> MessageBroker<C> {
 
     /// Send an input message to the component.
     pub fn send(&self, input: C::Input) {
-        self.inner.sender.send(input);
+        self.inner.sender.send(input).unwrap();
     }
 
     pub(super) fn get_channel(&self) -> (Sender<C::Input>, Option<Receiver<C::Input>>) {
         let inner = &self.inner;
-        (inner.sender.clone(), inner.input_rx.lock().unwrap().take())
+        (
+            inner.sender.clone(),
+            inner.input_receiver.lock().unwrap().take(),
+        )
     }
 }
 
 struct MessageBrokerInner<C: Component> {
     sender: Sender<C::Input>,
-    input_rx: Mutex<Option<Receiver<C::Input>>>,
+    input_receiver: Mutex<Option<Receiver<C::Input>>>,
 }
 
 impl<C: Component> MessageBrokerInner<C> {
     fn new() -> Self {
         // Used for all events to be processed by this component's internal service.
-        let (sender, input_rx) = crate::channel::<C::Input>();
+        let (sender, input_receiver) = crate::channel::<C::Input>();
         Self {
             sender,
-            input_rx: Mutex::new(Some(input_rx)),
+            input_receiver: Mutex::new(Some(input_receiver)),
         }
     }
 }